requirements.txt
SpeechRecognition==3.10.3
pyaudio>=0.2.14
pyttsx3==2.90
google-generativeai==0.4.1
python-dotenv==1.0.1
